% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weewx.R
\name{export_weewx}
\alias{export_weewx}
\title{Export WeeWX Data}
\usage{
export_weewx(db_path, vars = c("outTemp"))
}
\arguments{
\item{db_path}{Path to the WeeWX database file.}
}
\value{
A data frame with selected variables.
}
\description{
Export selected variables from a WeeWX database.
}
