% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weewx.R
\name{tidy_weewx_names}
\alias{tidy_weewx_names}
\title{Tidy WeeWX Variable Names}
\usage{
tidy_weewx_names(data)
}
\arguments{
\item{data}{A tibble containing the WeeWX data.}
}
\value{
A tibble with variable names converted to snake case.
}
\description{
Rename variables from WeeWX format to snake case.
}
